requirements.txt
torch
torchvision
numpy
tqdm
einops
opencv-python
torch-fidelity
Pillow
tensorboard
matplotlib
 